requirements.txt
keras == 2.2.4
n2 == 0.1.4
numpy == 1.16.4
pandas == 0.24.2
scikit_learn == 0.21.2
six == 1.12.0
tensorflow_gpu == 1.12.2
tqdm == 4.32.2
keras_transformer == 0.28.0
